Create policy for Provider

+++ app/Http/Controllers/ProviderController.php
@@ -18,6 +18,8 @@ class ProviderController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Provider::class);
+
         $pageSize = (int) $request->input('pageSize') ?? env('API_ITEMS_PER_PAGE');
         $page = (int) $request->input('page') ?? 1;
 
@@ -34,21 +36,25 @@ class ProviderController extends Controller
 
     public function store(StoreRequest $request)
     {
+        $this->authorize('create', Provider::class);
         return $this->service->store( $request, $request->validated() );
     }
 
     public function show(int $id)
     {
+        $this->authorize('view', Provider::class);
         return Provider::findOrFail($id);
     }
 
     public function update(UpdateRequest $request, string $id)
     {
+        $this->authorize('update', Provider::class);
         return $this->service->update( $request->validated(), $id );
     }
 
     public function destroy(int $id)
     {
+        $this->authorize('delete', Provider::class);
         if ( $provider = Provider::find($id) ) {
             $provider->delete();
         };
